"1:45pm. Done with breakfast. Let me do the chores and then I will resume my fight to get the Ocaml tooling working.
2:15pm. I am done. Let me finally resume. I'll leave my pleasures behind in order to do some research.

Ok, first of all. Can I use opam directly from Windows?

The reason why I need a Cygwin shell is because the Ocaml compiler uses Linux tools. So the answer would probably be no.

Still, let me just check what happens when I run it. After that I'll see how I can get the shell to start properly.

https://stackoverflow.com/questions/9637601/open-cygwin-at-a-specific-folder

2:30pm. That thing is to add a launch icon during right click. I have a different problem, namely that the Ocaml files do not work properly when I start in anything but the home folder.

...Ah, let me just ask.

2:50pm. https://stackoverflow.com/questions/60775511/how-to-stop-the-cygwin-terminal-from-forgetting-the-environment-when-opening-it

Finally, composed this.

https://stackoverflow.com/questions/23666173/opening-files-from-cygwin-home-in-windows-applications
"You can also install a Cygwin version of Emacs"

I can? That will make it a lot easier to set up.

3pm. https://www.cs.umd.edu/class/spring2018/cmsc330/ocamlInstallationGuide.pdf

I had not realized it at first, but this guide is a WSL guide.

If I backtrack and go down this route, will that make tooling integration easier? Will I be able to use the Ocaml tools directly? I do not really understand the pros and cons of WSL at the moment.

3:35pm. Damn, how time consuming this is. Enough Reddit posting. I'll see if I get any good replies by the end of the day. I did get a good hint by that guy who told me the `reason-vscode` plugin supports both Reason and Ocaml.

If I do not get anything more, I'll just ask the author for advice.

For now, let me resume reading the documentation. In the worst case, I'll just use Emacs for Cygwin in order to get Ocaml to work.

3:45pm. Let me resume the documentation. I forgot, where was I before?

```
a(one, two, three)
```

Is the same as...

```
one->a(two, three)
```

Uf, this definitely is not the usual pipe operator behavior like in F#.

Ah, it does seem that `|>` exists and does what I would expect. This here is a special JS thing for objects because JS methods are the wrong way around.

3:55pm. "A variant's constructors, like Some or Student, look like functions, but unfortunately aren't, due to historical reasons. Sometime, it'd still be nice to be able to use them as functions. Pipe First takes the occasion to enable that for you!"

In Ocaml they aren't constructors? That is interesting.

4pm. "The principle of least power applies when you're trying to "Get Things Done". If the problem domain allows, definitely pick the least abstract (aka, the most concrete) solution available, so that the solution is reached faster and has fewer unstable indirections you'd have to traverse. For example, prefer types over free-form data, prefer data-driven configuration over turing-complete function calls, prefer function calls over macros, prefer macros over project forks, etc. When you constrain your domain and power, things become easier to analyze. That is, if the domain is constrained enough to allow it."

The documentation says good things. Incidentally, Haskell violates this constantly.

"Sometimes, it's fine to write a little bit of boilerplate to reduce the need for otherwise extra powerful types. If anything, tasteful tradeoffs might show your pragmatism and judgement more than fancy types!"

This is true too.

4:05pm. Now I am looking the VS Code plugin again.

"This project implements the Language Server Protocol for Reason.

It is written in Reason, and compiled w/ esy + dune. The goal is for it to work equally well on Windows, MacOS, and Linux."

"(but not under WSL, make sure you install bs-platform etc. from the windows side)"

I am not sure that the project supports Ocaml after all.

But I find it interesting that it is written in Reason itself. That is some good info on me. It will make studying it easier.

4:10pm. Well, forget that. Let me resume the Reason documentation. I'll aim to finish it all today.

This stuff is easy and I am completely familiar with it.

4:30pm. https://reasonml.github.io/docs/en/exception

"The above tip seems to contradict what's happening in the OCaml standard library; prominent functions in modules such as List and String seems to throw exceptions overly often. This is partially a historical sediment, and partially out of extreme care for performance. Native OCaml/Reason is incredibly performant; exception throwing was designed to be very cheap, cheaper than allocating and returning e.g. an option. This is unfortunately not the case for JavaScript."

The same goes for F# as well. I remember Jon Harrop talking about how expensive .NET exceptions are compared to Ocaml.

At any rate, a thought occurs to me. What if the ReasonML plugin is not using the compiler itself to do type inference? That might be doable in some cases.

Well, it is probably not the case here, but I should leave my mind open.

4:50pm. https://reasonml.github.io/docs/en/module

The stuff here is a pretty good explanation of modules. I actually hadn't known how they work till now.

I admit, the way objects work is fairly bizarre and I hadn't understood the difference between open and closed ones or what that `as 'a` is supposed to represent.

5:05pm. This was interesting. This documentation is good. I finally understand what modules are.

And finally done with the basics section. I'll finish the whole documentation soon.

5:10pm. https://reasonml.github.io/docs/en/interop

The way Bucklescript does interop is similar to how I will do it in JS.

By the way...Jon Harrop talked about modules being super complex in the various SML implementations, but from what I've seen, they aren't a big deal.

5:30pm. https://reasonml.github.io/docs/en/converting-from-js

I'll really focus on this for later.

https://reasonml.github.io/docs/en/native

Let me just take a moment to pause for a bit and catch my breath.

The reason why I am going for ReasonML is because I am actually going to use it. I am seriously considering replacing large parts of `vscode-languageserver-node`, but even though I said that, doing it in either TS or Purescript does not appeal to me. In Reason, I should do it in a proper style.

Now let me get back to the docs.

"Bsb-native is a fork of BuckleScript's bsb that compiles to native OCaml instead."

Why not just use Ocaml then?

...Well, besides the Windows issues?

"Note: Make sure you have the latest esy installed. If in doubt, run npm remove -g esy && npm install -g esy@latest."

I actually did not know about `npm remove`. I was using `uninstall` previously.

"To add native Reason packages that happen to be hosted on npm, run esy add npm-package-name."

This is pretty interesting. Could I do the same for Spiral packages eventually. The stuff Purescript does is interesting, but I want to take the line of least resistance when it comes to doing package management for my own language.

"esy treats the npm scope @opam specially. esy will install any package name with the @opam scope directly from opam. This is the only scope with special meaning. All other package names are assumed to be hosted on npm."

So I can use `opam` using `esy`?

5:40pm. https://reasonml.github.io/docs/en/project-structure

Just this thing and then I am done, possibly for the day.

After I finish, my priority will be to check out the Reason plugin. Today I put some effort into Ocaml interop, but now that I know it is entirely written in Reason, I do not have to hesitate in checking it out.

5:55pm. Done with the Reason documentation.

I'll call it a day here.

Still haven't gotten a reply to that SO question. "