"1:25pm. Done with breakfast. Let me do a little update to that issue that I've opened. I've given so little info to the author that I am afraid it will simply get ignored.
One of the problems is that the plugin is trying to open the compiled server. I might be able to run the project properly if I can instruct the plugin to ignore that. There is an option in the plugin settings, but I do not know how to use it.

1:35pm. "Before I built the project, I had a similar error to the one above when I opened the project folder in VS Code. Back then as part of the error there was a path a missing bin.exe file somewhere in the project. I think the project built successfully, but the plugin itself might be doing something wrong.

Until this gets resolved would it be possible to redirect the plugin so it uses the server bundled with it rather than the one being compiled here? Right now I just want to browse this project and want editor support to go along with it."

This is good. This should be more constructive. The issue is number 426 on the `reason-language-server` repo. I do not want to paste the link here directly as it will get referenced. In the past I've made this embarrassing mistake and have my journal entries as part of commits.

Now, the next thing on my list would be chores. I have a big pile of lumber I need to put into storage and it will take me an hour extra every day until that is done. I should not take too long.

1:45pm. Let me turn off here for a while. I've really been going full steam even though I said I would be resting."