"4:25pm. I am back. Let me do some chilling.
https://www.reddit.com/r/MachineLearning/comments/flqzsb/d_the_end_of_starsky_robotics/
https://www.reddit.com/r/MachineLearning/comments/flja88/d_is_most_deep_learning_research_is_useless/

4:50pm. I suppose that was good enough to cool me down a bit.

Now, let me do the last bit of work for the day.

I need to focus on the next part.

How does the lsp-sample interface with VS Code itself. All this time my attention was on the editor, but if you want to pull something out, do it by the roots.

```
  let diagnostic: Diagnostic = {
   severity: DiagnosticSeverity.Warning,
   range: {
    start: textDocument.positionAt(m.index),
    end: textDocument.positionAt(m.index + m[0].length)
   },
   message: `${m[0]} is all uppercase.`,
   source: 'ex'
  };
```

Unfortunately, the textDocument stuff is from the library itself.

5:10pm. Had to take a break. Let me resume.

The above is actually a big insight for me.

It did occur to me to wonder why didn't I figure out where the editor boundary is last time, and this makes it clear why I missed it.

5:15pm. It is really simple, but this kind of thinking is really key. Simply knowing where the boundaries are is a great advantage.

Let me hunt down `TextDocument` in the other project. That is what I should be studying.

5:20pm. Or at least, what I would if I weren't so distracted by thoughts of interfaces.

Yeah, interfaces are the way to go here. Being able to associate methods to nominal types would go a significant way towards resolving the design issues with type classes even though the approach is slightly less powerful.

5:25pm. I am thinking about it, and the general approach seems like it will be a lot easier to evolve. It is a lot more robust.

5:35pm. Yeah, I am warming up to this approach. I'll allow type constraints just for nominal types. All in all, this is a very similar design to F# except I will have HKTs.

5:40pm. ...There are some issues unfortunately.

No, this approach is not good enough. If I was doing bottom-up only, then it would be ideal, but those damn HTKs...

I do not know how to unify them when the number of type args in the interface and the number of type args in the nominal do not match up.

5:45pm. Agh, damn it. I am too tired for this. Let me check out that `TextDocument`.

Ah, the thing has its own project.

5:55pm.

```
function mergeSort<T>(data: T[], compare: (a: T, b: T) => number): T[] {
 if (data.length <= 1) {
  // sorted
  return data;
 }
 const p = (data.length / 2) | 0;
 const left = data.slice(0, p);
 const right = data.slice(p);

 mergeSort(left, compare);
 mergeSort(right, compare);

 let leftIdx = 0;
 let rightIdx = 0;
 let i = 0;
 while (leftIdx < left.length && rightIdx < right.length) {
  let ret = compare(left[leftIdx], right[rightIdx]);
  if (ret <= 0) {
   // smaller_equal -> take left to preserve order
   data[i++] = left[leftIdx++];
  } else {
   // greater -> take right
   data[i++] = right[rightIdx++];
  }
 }
 while (leftIdx < left.length) {
  data[i++] = left[leftIdx++];
 }
 while (rightIdx < right.length) {
  data[i++] = right[rightIdx++];
 }
 return data;
}
```

Implementing merge sort by hand. That is a bad look.

These kinds of algorithms are specifically ones that you'd be better off just importing. It is amazing how JS programmers import 100s of useless dependencies without blinking an eye and yet they have no problem reimplementing something that should be in the standard library.

6:05pm. Actually, I think my idea would work. Yeah, there are some edge cases, but there are ways of working them out.

...And for all my kwetching, I think that the stuff in `TextDocument` is actually important. If I were to reimplement something in Reaosn, this would definitely be top priority.

6:15pm. Actually more than just `TextDocument`, in the server there is something called `TextDocuments`. That thing handles text document events.

That is what I need. That is one of the roots which no doubt interfaces with the editor...

But aren't I forgetting one thing - where are the text documents being passed to it?

```
// Make the text document manager listen on the connection
// for open, change and close text document events
documents.listen(connection);
```

Via the `connection`. Hmmm...really?

6:25pm. Ah, finally I found it on line 156 of `main.ts` in `server`.

6:30pm.

```
  (<ConnectionState><any>connection).__textDocumentSync = TextDocumentSyncKind.Full;
  connection.onDidOpenTextDocument((event: DidOpenTextDocumentParams) => {
   let td = event.textDocument;

   let document = this._configuration.create(td.uri, td.languageId, td.version, td.text);

   this._documents[td.uri] = document;
   let toFire = Object.freeze({ document });
   this._onDidOpen.fire(toFire);
   this._onDidChangeContent.fire(toFire);
  });
```

Because of stuff like this, I can never tell whether the types make sense for what I am looking at.

6:45pm. I am still obsessed about typeclasses. Actually, I just realized something. There isn't a situation where I really need more than argument in them.

7:05pm. No the interface view is not working out in my mind.

7:10pm. I am completely perplexed. Forget this for the time being. I was not in the right state of mind to think about to begin with.

```
public listen(connection: IConnection): void {
```

At any rate, at least now I have a lead what I need to look for.

I did the bottom part with `jsonrpc`, but now it is time for the big leagues. I've gotten a decent bit familiar with Typescript's style.

7:20pm. Actually, the inspiration is coming to me. Actually the interface view does make sense. I know how to do it.

I think that now the connection between typeclasses and OO classes is completely clear to me.

Let me call it a day here."