"2:25pm. Done with breakfast. Even though it got colder, it seems today the rain did not come as expected so I'll be outside doing the chores for a while longer again.
Let me chill for a bit first.

2:30pm. Today is strange. I never really know when I'll go into thinking mode. But that I have the leisure to seriously think about extensibility means that I am getting closer to finishing the phase I am currently on even if it does not seem like it. My subconscious would not have guided me in this direction if it did not have the mental resources to do this.

3pm. Oh yeah, this idea occurred to me earlier. Before I leave for the chores, let me just take a look at how Scala does monads.

The way Spiral works is well suited to OO after all.

https://medium.com/free-code-camp/demystifying-the-monad-in-scala-cc716bb6f534

```
trait M[A] {
  def flatMap[B](f: A => M[B]): M[B]
}

def unit[A](x: A): M[A]
```

So something like...

```
interface type bind m a : m a -> forall b. (a -> m b) -> m b
```

I am still not decided on syntax, but the idea I had is quite similar to scala isn't it?

It might be a good idea to look into how Scala does it before I decide.

```
def unit[A](x: A): M[A]
```

The thing I am most unsure is this. This is not like Haskell so I can't do a typeclass search. Just how does Scala do it?

3:10pm. I want to do the chores so let me take off here. I still do not know the semantics of how Scala does it, but I will leave that for later.

Hopefully tomorrow I will get an answer from `reason-vscode` author so I can start studying it. If I don't I am going to have to figure out how to deal with the issue on my own. I am not sure if I'll be to do that. I'll see how things go.

Either way, I'll finish the interface design for Spiral soon and get a move on with that."