"10:45am. I had a lot trouble falling asleep, but somehow once I did it was deep, so I feel quite refreshed right now.
10:50am. I feel like my focus is going up. I've moved on from thinking about Spiral's type interfaces, to thinking about reactive programming. I've been thinking about Svelte and contrasting it with my 2016 experience making that GUI poker game.

10:55am. Up to now I've never really understood the usefulness of interfaces, but I think I finally see it now.

11am. Back then I had no idea what to cut, but I really should have put the state at the top and the controls at the bottom, but I could not really see it.

11:25am. Had to the chores early today. Let me resume.

Yeah, having the controls at the top and having the game directly mutate them is nothing more than spew. Even if I could not have written it directly in a language specialized for it, when programming you are either abstracting or you aren't, and I definitely wasn't at that time.

It is one of my more embarrassing programming experiences.

I've been looking for a way to get payback for a long time now.

11:35am. The reason why I am talking about this now is because there is not much difference between GUI programming and doing editor support. It is the same reactive programming challenge that I completely failed last time, and for it in particular I won't be able to use Svelte for this. I have to level up on my own here.

There has to be a reason why every plugin is such a mess, and no doubt it is the same reason why that GUI I did ended up being 1k LOC rather than 50-100.

11:40am. Still no reply to the Reason plugin issue. That thing is not going to go anywhere as expected.

I am 100% sure now - I need to focus on the VS Code project. I am going to finish studying it and then start coding.

I've done quite a bit of study, and I think it is time to start winding down on that. Just like I've been in overdrive regarding how to do the typeclasses and row polymorphism in Spiral, it is time to directing that energy towards editor support. My mental rehearsing is already moving in that direction.

11:45am. Let me have breakfast here. Blattodea and a lot of other stuff is out. I'll start for real later."