"8:45am. I somehow managed to pull my calf muscle during the night so I'll be confined to this chair for the next few days...
Much like with the recent Corona events, this will not noticeably impact my lifestyle.

Yesterday I got Doom Eternal and it is very fun. I am so bad at it, and I am still making good progress on Nightmare. The sheer amount of combos in this edition is somewhat bewildering, and I am yet to get a grasp on it.

8:50am. Well, let me do my morning reading and then I will start.

9:25am. Ok, let me just read the latest Astrid chapter (it is really taking too long to get to the genocide) and then I will start.

9:35am. Let me finally start.

What I will do first is go through the examples for chapter 5.

9:55am. Had to take a break.

I came to a realization. If I hurry I could do it in 2 days, but I want to take my sweet time with this in order to study the examples properly. 2 days is just not enough time. I feel too much rushed right now.

Forget my plan of delaying the review tomorrow. I will do it normally, and then give the book the respect it deserves after that.

10:05am. Focus me, focus. Study the examples, not /pol/ threads.

```
    class VariableTimePrimeCheckService : PrimeCheckService
    {
        private readonly int _numberToDelay;

        public VariableTimePrimeCheckService(int numberToDelay)
        {
            _numberToDelay = numberToDelay;
        }
```

I always taken aback at how poorly the mainstream OO languages are designed. Here is the F# version.

```
type VariableTimePrimeCheckService(numberToDelay : int) = ...
```

All they had to do was put the fucking argument at the top, and yet they could not do that much. It is amazing. It is like zero thought went into the ergonomics of the language. This is what happen when you let NPCs do the design.

10:25am.

```
        [<CompiledName("ZeroCreate")>]
        let zeroCreate count =
            if count < 0 then invalidArgInputMustBeNonNegative "count" count
            Microsoft.FSharp.Primitives.Basics.Array.zeroCreateUnchecked count

        [<CompiledName("Create")>]
        let create (count: int) (value: 'T) =
            if count < 0 then invalidArgInputMustBeNonNegative "count" count
            let array: 'T[] = Microsoft.FSharp.Primitives.Basics.Array.zeroCreateUnchecked count
            for i = 0 to Operators.Checked.(-) array.Length 1 do // use checked arithmetic here to satisfy FxCop
                array.[i] <- value
            array
```

I know that I should be studying RP, but I am thinking of the issue of serializing types in Spiral. Remember that union type to dense vector translation. That kind of thing.

Trying to figure out the right way to design `tensor_init` has been giving me trouble because init requires a size and a function to go along. All this time I've been thinking that just the type would not be enough because in .NET arrays are always zeroed out - so by that logic if I go with `init` I'd get an efficiency gain. But no, creating them raw is what happens in the F# standard library. As one might expect.

Well, it is one of the issues that was concerning me. For real nominal types, I need to take some special care to make sure that one can go from types to values for the sake of closure conversion.

I'll make the syntax something like...

```
real nominal Tensor l a : (i : l) = tensor_create `a i
```

With this, the partial evaluator will have enough information to make both the type and provide its constructor arguments. I've been looking for ways to avoid having to make separate constructor and type provider code and this will do the trick.

Yes, I know that having the constructor be a part of the type was my plan from the start, but I've also been considering other doing the type provider separately.

Incidentally, having to do write separate type inference code in Spiral by hand is why I am putting global type inference in v0.2. This is something to be avoid if at all possible.

10:50am. I am thinking of how record application could be done in the typed segment.

I need that to do tensors with named dimensions.

10:55am. Ah, enough of this. I will find a way to do record application with the power of constraints. It will definitely be possible.

Let me get back to studying the examples.

This is what I meant by needing to take my time. I really do need time to think in order to establish the foundations.

11am. Oh, there is a `CreatingPeriodicUpdatableView` WPF project in chapter 5. It had not been mentioned in the book. It is really quite nice of the author to this. If I was writing a review I'd give it an extra point just for this.

11:15am. I really wonder how the C# compiler knows to go after `app.xaml`. Where is this set? At some point I am going to have to ask around, but nevermind that for now. XAML is decently readable.

11:25am.

```
            foreach (var prime in await generator.GenerateAsync(5))
            {
                Console.Write("{0}, ", prime);
            }
```

Hmmm, you can put `await` inside `foreach`? This is interesting. C# is neat in weird places.

11:45am.

```
        private static void UsingThreadPool()
        {
            Console.WriteLine();
            Demo.DisplayHeader("Using the ThreadPool");

            Console.WriteLine("Starting the long work");
            ThreadPool.QueueUserWorkItem((_) =>
            {
                //simulating a long work of five secnods
                Thread.Sleep(TimeSpan.FromSeconds(5));

                Console.WriteLine("Long work is done");
            });

            Console.WriteLine("Waiting for long work to finish, but you can press <Enter> any time to continue");
            Console.ReadLine();
        }
```

So you can use the thread pool like this.

I am really taking my sweet time here, but studying should be done at a comfortable pace as it is a creative endeavor.

11:50am. Let me finally start chapter 6. I'll read just a few pages and then stop for breakfast. This book has been pretty good to me so far.

12:15pm. "To capture and handle all the unhandled exceptions thrown from tasks, you can use the TaskScheduler.UnobservedTaskException event that will be triggered when a task is disposed of because its exception wasn’t observed. You can also change the default behavior so that the process will terminate by setting a configuration in your app.config\web.config file."

I don't understand how this works.

https://docs.microsoft.com/en-us/dotnet/api/system.threading.tasks.taskscheduler.unobservedtaskexception?redirectedfrom=MSDN&view=netframework-4.8

"This application domain-wide event provides a mechanism to prevent exception escalation policy (which, by default, terminates the process) from triggering.

To make it easier for developers to write asynchronous code based on tasks, the .NET Framework 4.5 changes the default exception behavior for unobserved exceptions. Although unobserved exceptions still raise the UnobservedTaskException exception, the process does not terminate by default. "

Ah, I think I see it in terms of individual tasks at least. I am not sure how the observable would know whether the exception is unobserved...though it would be implementable by letting the exception flow to the end and reraising that event.

I am just guessing here.

12:25pm. 155/337. Let me stop here."