"10am. I am up. Today is review time. Let me chill a bit and then I will do it.
10:55am. Enough chilling, let me start.

Review:

---

For me, March was a continuation of February. I've been trying to figure out a way to do editor support for Spiral properly and have been doing thing related to that like getting familiar with webdev and studying other people's code and language plugins. I learned some [useful basics](https://www.youtube.com/playlist?list=PLillGF-RfqbYeckUaD1z6nviTp31GLTH8) and got familiar with various web languages. After all of that, my conclussion is that if you want to know how to do editor support for your own language then studying how other people have done it is a waste of time. Last week it occured to me that the reason the quality of the code I've been running into is so low is quite similar to the reason why the one time I did a GUI back in WPF back in 2016 turned into such a huge mess.

I've always had great respect for tools that had good UIs. It was in late 2015 that I made F# my main language because how lightweight it felt and how good its type inference was. Before that, I tried Scala for the FP Coursera course and never got the same ease from it. I actually took F#'s editor support for granted and ended up being greatly dissappointed when I tried Haskell.

Last year, I tried Pharo, but it was not quite what I was looking for. I received an epiphany when I discovered [Svelte](https://svelte.dev/). If I were asked at what is high quality code, I'd point at those examples. The form and the concision is perfect, and the realization that the possibility exists made me think of what exactly I was missing during my WPF days, my Pharo days and my current language plugin days. It lingered in the back of my mind as I read through the language projects, putting in great effort, but absorbing little in the process.

Back in 2015, after doing the intro to FP course in Scala, I did the reactive programming one. I wish I could say it inspired me. Even though it introduced me to async/await, [reactive extensions](http://introtorx.com/) and message passing concurrency, I did not attain any true understanding of the subject. Later, when the time came to apply those skills to UIs I never made use of what I have learned and code monkeyed it instead to disappointing results.

Though I've learned much since that time, I am still half baked as a programmer. There is no way around it - most programs that have real world value have a graphical aspect to them and inolve user interfaces. Or like editor support for languages, have an aspect of concurrency to them similar to UIs.

One way to get the required productivity is to study languages like Svelte, or pick up a framework, but that is not the answer. As nice as Svelte is - the general model of it is too restricted. You can't do editor support for a language in Svelte or some framework. It leans on the code monkey side of the value curve; knowing those things is not a fundamental programming skill like the understanding of combinator libraries is.

A programmer must by necessity make value judgements, and if he mistakenly rates fundamental skills as being of low import then he is going to find it very hard to get better. The substitutes for fundamental skills are always second rate.

The answer of what I need to do here is simple - master functional reactive programming.

Editor support for Spiral should not be a big deal once I level this up, so I'll forget about that for a while and study this properly. From here on out and without limit, this skill is one that I will be using everywhere. For the sake of grasping reality, I will change my programming.

---

12:35pm. https://www.youtube.com/watch?v=CFW45QqtXKE&list=PLB71qIrDRO8ifVSUz537dZFjYQ-jyIs5w&index=32
"Dr.Hayden : noooo you cant just
shoot a hole into the surface of mars

Doomslayer : haha big gun go boom"

Let me stop here so I can have breakfast. I'll get back to the book after that. As it turns out, I never had the chance to mention it in the review."