"10:15am. Yesterday's session was great. But now, let me see if I can something done. First the review. The one I wrote was too meandering. Let me redo it.
Review:

---

Plan A: study other people's language plugins - waste of time.

Plan B: master [functional reactive programming](http://introtorx.com/) - in progress (1 week.)

I tend to do things in two ways - either do things in the least amount of effort, or if that is impossible do it in the most amount of effort in order to build the foundations. After wasting 4 out of the last 7 weeks, I eventually realized that me being bad at things like UIs and me not having an alternative to code monkeying editor support are all connected.

I really should have mastered FRP long ago, but the disadvantage of learning it all on your own is that the order you tackle the tasks in might turn out to be suboptimal and I've gotten bit by that here. This will take a few months, but hopefully not more than that. One benefit is that this is giving me time to get the design of Spiral v0.2 all in order before I start.

Without a doubt FRP is the essential next step from working with events. If it were just UIs, there are various frameworks you can make use of depending on the platform, but in my mind those lean more on the code monkey side of things. You cannot use them to abstract editor support for a language, but FRP is different.

Acquiring this sort of skill has great appeal to me. Most programs worth using in the real world have a graphical user interface, and reactive components to them. After I learn this, I will finally be able to graduate from making command line tools and move to making truly high quality ones with great proficiency.

By the end of all of this not only will I have Spiral v0.2, but I will have all the skills one would expect a professional programmer (the kind that makes money) to possess. I am looking forward to that.

---

This is perfect.

10:55am. Yeah, this is good enough to post. It is not a rambling sequence like the last review.

11am. Done with that.

11:05am. Let me go to the next thing. IntroToRx.

11:10am. The stuff in it is much like in the book I just read. Nevermind that. I am done with this. I am done with FRP basics.

The next comes the Akka book. I want to have so fun with that. I do not need to study message passing concurrency, but I might as well do it since I am at it.

11:20am. The Java book I got on a whim looks pretty good, but it is of no use to me. I guess I'll keep it on my drive.

Let me read the first chapter of the Akka book.

11:45am. 16/264. Done with chapter one and got the code samples. Let me stop here.

1:30pm. Let me read a chap of Overlord, and then I will start.

1:55pm. Had to do the chores. The chapter is next.

2:10pm. Let me start. I've gone through chapter one of the Akka book, so let me check out the samples.

Since I forgot to commit earlier, let me do it now."