"10:30am. As great as Doom Eternal is, I am going to have to curtail my gaming activities a little as it seems I am taking the time out from my work hours from the other end.
10:55am. Let me start. Yesterday a video caught my eye.

https://www.youtube.com/watch?v=ZCRYBivH9BM
Fabulous – F# for cross-platform mobile apps - Don Syme

Let me start the day with this. I'll do the tutorials later.

11:10am. https://youtu.be/ZCRYBivH9BM?t=807
"...Revaluating the view on changes to the model. And suddenly all these warning signs go off in your head - isn't that going to flicker, isn't that going to be too costly. Isn't that...can't thinking flow back?"

11:40pm. https://youtu.be/ZCRYBivH9BM?t=2277
"The model has a DoNotUseMe type. You cannot actually use it internally."

Oh so depends on is not as unsafe as I thought. Interesting.

11:45am. https://docs.microsoft.com/en-us/xamarin/get-started/what-is-xamarin-forms

This talk is making me wonder whether Xamarin Forms can be used on Windows. I thought it was a mobile thing.

> Xamarin.Forms is an open-source UI framework. Xamarin.Forms allows developers to build Android, iOS, and Windows applications from a single shared codebase.

That answers that.

Why don't I seriously consider XF for once? I mean since Fabulous works nicely with this, why not? It is not like I am married to WPF.

https://youtu.be/ZCRYBivH9BM?t=2428

So it has charing, maps, and animations. Everything I'd need for a poker app.

https://youtu.be/ZCRYBivH9BM?t=2597

He is saying that F# is doing some special compiler optimizations to make this go fast. Actually, it is really good that Don Syme is working on this. UIs are one place where language support is necessary.

12:30pm. Did a little reply to that guy in the PL monthly thread.

https://gist.github.com/staltz/868e7e9bc2a7b8c1f754
The introduction to Reactive Programming you've been missing

This is by the cycle.js author. I'll give it a read later.

Let me have breakfast here. I want to watch more talks on this. I think I'll go for the ones by Jim Bennet and Thomas Petricek next. After that I'll give Elmist.WPF a try. Then comes Fabulous.

I'll play with this for a while. Since these libraries will make doing UIs much easier, I'll indulge myself with this here."