"12:45pm. Done with breakfast. The next come chores.
But let me slack a bit first. What I will do when it is time to resume is go through that cycle.js course first. I want to get the easy stuff out of the way.

https://www.maa.org/external_archive/devlin/LockhartsLament.pdf
A Mathematician’s Lament

I'll read this later.

1:35pm. Done with chores. Let me read that thing and go through the course after that.

1:45pm.

> Worse, the perpetuation of this “pseudo-mathematics,” this emphasis on the accurate yet mindless manipulation of symbols, creates its own culture and its own set of values.

> And I haven’t even mentioned the lack of mathematical criticism in school. At no time are
students let in on the secret that mathematics, like any literature, is created by human beings for
their own amusement; that works of mathematics are subject to critical appraisal; that one can
have and develop mathematical taste.

The paper is really good.

Mathematical taste is something I only really got in 2019 thanks to my type theory effort.

2:45pm. This took me a while. Let me go through that course. To be honest, I am bored with just reading and studying, the design work I did earlier was much more interesting in comparison, but let do things according to plan.

With regards to `cycle.js`, the talk I watched was from 2015 and I want to see what the guy came up with in the newer versions.

2:55pm. https://egghead.io/lessons/rxjs-separate-logic-from-effects-in-cycle-js

Oh, rather than talking about `cycle.js`, he is actually going to build it. That is great. I expected this course to be a snoozefest, but you can tell by just the first 4m that it will be worth watching.

3:10pm. https://egghead.io/lessons/rxjs-read-effects-from-the-dom-with-click-events

Remember that problem I talked about yesterday of not understanding where the first step comes from. He tackles it right here.

3:15pm. It was roughly what I thought it would be, but what is the fake thing though? Wouldn't doing things like this run into issues?

Well, I suppose it work for him, but I'd much prefer the message pump version. Once this is done what I am going to do is ask around whether a library like I have in mind has already been done. Maybe I'll do a little proof of concept if not.

3:35pm. https://egghead.io/lessons/rxjs-fine-grained-control-over-the-dom-source-in-cycle-js

I admit I am not paying as much attention as I should since I am thinking about how to do my own library, but this is a really cool video.

3:45pm. I am falling into daydreams. I really am getting excited at the possibility of that proof of concept Rx UI library. I really want to do it just to check that it is possible and to show off a little.

3:50pm. Forget it for now. Let me just watch the videos and then I'll think about seriously doing it. It might be a good first project to refresh my memory of WPF.

3:55pm. https://egghead.io/lessons/rxjs-fine-grained-control-over-the-dom-source-in-cycle-js

I am reminded of that HN when one of the users raised issues with the component state. This would definitely be a problem with this sort of arrangement.

4:10pm. https://egghead.io/lessons/rxjs-read-text-input-in-a-cycle-js-hello-world-application

This guy is definitely a real programmer. He gets it. Though he did the wrong design.

4:20pm. https://egghead.io/lessons/rxjs-read-text-input-in-a-cycle-js-hello-world-application

It does need to be initialized with some state as expected.

5:10pm. http://codebetter.com/matthewpodwysocki/2010/02/05/using-and-abusing-the-f-dynamic-lookup-operator/

Since I do not want to write boilerplate for what is just a proof of concept, I'll refresh my memory of F#'s dynamic features and use those. It should be nice.

https://egghead.io/lessons/cycle-js-use-multiple-independent-instances-of-a-cycle-js-component

I can do better than this. I can make a library that combines all the benefits of Fabulous and having access to reactive extensions.

5:20pm. https://egghead.io/lessons/cycle-js-isolate-component-instances-in-cycle-js

I am not listening at all by this point as I do not like this, but since there is only a little bit left let me finish it. After that I'll call it a day. I definitely won't be pushing myself till 8pm like yesterday.

5:25pm. Let me do this, let me finish this. Then I'll ask online about doing the MVU pattern using reactive extensions.

5:40pm. Can't find anything with Google on the MVU pattern using reactive extensions. Let me ask on the PL sub then.

6pm. https://www.reddit.com/r/ProgrammingLanguages/comments/fw1944/are_there_any_libraries_that_do_the_mvu_elm/
"It is used to do UIs. Here is [an example](https://fsprojects.github.io/Fabulous/Fabulous.XamarinForms/#a-basic-example). Also known as the Elm pattern, it was pioneered by Elm and is now being picked up by other languages in the form of libraries.

I've been studying both this and [reactive extensions](http://reactivex.io/intro.html), and I've realized that it would not be hard at all to implement this pattern by via compilation to a chain of observables of UI components. This would have the advantage of not needing view diffing and having access to the full power of reactive extensions at every point in the program.

Since I've spent time planning it out, I'll do a proof of concept of this if nobody else has done it. Google is not giving me anything, and some of the reactive JS DOM libraries I've looked at like `bacon.js` and `cycle.js` have different approaches."

I'll see what replies I'll get.

Right now, I feel like calling it a day. Today I've been low energy; I haven't really done much except dwell on this idea all day. I am really happy I came up with it. I feel that at this point I finally have a handle on UIs. This has been something I've been trying to master for years and I finally know the things I've been missing.

What I have now are two goals only:

1) Study Rx.NET's source. Find out how it is implemented in depth and what those other things in the project are.
2) Work on the MVU via Rx proof of concept library.

I'll give this thing a while to settle.

After I've cleared these two hurdles, I'll be ready to start work on editor support for Spiral. You can bet that I'll find a use for Rx there.

No monkeying this time around."